carbgai/requirements.txt
fastapi
uvicorn[standard]
python-multipart
transformers
torch
torchvision
Pillow
numpy
opencv-python
